refactor file structures: splitting the code into multiple files to optimize for readability
