--- go/tasks/plugins/k8s/sagemaker/sagemaker.go
@@ -1,833 +1,0 @@
-package sagemaker
-
-import (
-	"context"
-	"fmt"
-	"strconv"
-	"strings"
-	"time"
-
-	awsUtils "github.com/lyft/flyteplugins/go/tasks/plugins/awsutils"
-
-	hpojobController "github.com/aws/amazon-sagemaker-operator-for-k8s/controllers/hyperparametertuningjob"
-	trainingjobController "github.com/aws/amazon-sagemaker-operator-for-k8s/controllers/trainingjob"
-	"github.com/lyft/flytestdlib/logger"
-	"github.com/pkg/errors"
-	"k8s.io/client-go/kubernetes/scheme"
-
-	"github.com/lyft/flyteplugins/go/tasks/pluginmachinery"
-	"github.com/lyft/flyteplugins/go/tasks/pluginmachinery/ioutils"
-
-	flyteIdlCore "github.com/lyft/flyteidl/gen/pb-go/flyteidl/core"
-
-	pluginsCore "github.com/lyft/flyteplugins/go/tasks/pluginmachinery/core"
-	"github.com/lyft/flyteplugins/go/tasks/pluginmachinery/k8s"
-	"github.com/lyft/flyteplugins/go/tasks/pluginmachinery/utils"
-
-	commonv1 "github.com/aws/amazon-sagemaker-operator-for-k8s/api/v1/common"
-	hpojobv1 "github.com/aws/amazon-sagemaker-operator-for-k8s/api/v1/hyperparametertuningjob"
-	trainingjobv1 "github.com/aws/amazon-sagemaker-operator-for-k8s/api/v1/trainingjob"
-	"github.com/aws/aws-sdk-go/service/sagemaker"
-
-	taskError "github.com/lyft/flyteplugins/go/tasks/errors"
-
-	flyteSageMakerIdl "github.com/lyft/flyteidl/gen/pb-go/flyteidl/plugins/sagemaker"
-
-	"github.com/lyft/flyteplugins/go/tasks/plugins/k8s/sagemaker/config"
-)
-
-// Sanity test that the plugin implements method of k8s.Plugin
-var _ k8s.Plugin = awsSagemakerPlugin{}
-
-type awsSagemakerPlugin struct {
-	TaskType pluginsCore.TaskType
-}
-
-func (m awsSagemakerPlugin) BuildIdentityResource(_ context.Context, _ pluginsCore.TaskExecutionMetadata) (k8s.Resource, error) {
-	if m.TaskType == trainingJobTaskType || m.TaskType == customTrainingJobTaskType {
-		return &trainingjobv1.TrainingJob{}, nil
-	}
-	if m.TaskType == hyperparameterTuningJobTaskType {
-		return &hpojobv1.HyperparameterTuningJob{}, nil
-	}
-	return nil, errors.Errorf("The sagemaker plugin is unable to build identity resource for an unknown task type [%v]", m.TaskType)
-}
-
-func (m awsSagemakerPlugin) checkIfRequiredInputLiteralsExist(inputLiterals map[string]*flyteIdlCore.Literal, inputKeys []string) error {
-	for _, inputKey := range inputKeys {
-		_, ok := inputLiterals[inputKey]
-		if !ok {
-			return errors.Errorf("Required input not specified: [%v]", inputKey)
-		}
-	}
-	return nil
-}
-
-func (m awsSagemakerPlugin) BuildResourceForTrainingJob(
-	ctx context.Context, taskCtx pluginsCore.TaskExecutionContext) (k8s.Resource, error) {
-
-	logger.Infof(ctx, "Building a training job resource for task [%v]", taskCtx.TaskExecutionMetadata().GetTaskExecutionID().GetGeneratedName())
-	taskTemplate, err := getTaskTemplate(ctx, taskCtx)
-	if err != nil {
-		return nil, err
-	}
-
-	// Unmarshal the custom field of the task template back into the Hyperparameter Tuning Job struct generated in flyteidl
-	sagemakerTrainingJob := flyteSageMakerIdl.TrainingJob{}
-	err = utils.UnmarshalStruct(taskTemplate.GetCustom(), &sagemakerTrainingJob)
-	if err != nil {
-		return nil, errors.Wrapf(err, "invalid TrainingJob task specification: not able to unmarshal the custom field to [%s]", m.TaskType)
-	}
-	if sagemakerTrainingJob.GetTrainingJobResourceConfig() == nil {
-		return nil, errors.Errorf("Required field [TrainingJobResourceConfig] of the TrainingJob does not exist")
-	}
-	if sagemakerTrainingJob.GetAlgorithmSpecification() == nil {
-		return nil, errors.Errorf("Required field [AlgorithmSpecification] does not exist")
-	}
-	if sagemakerTrainingJob.GetAlgorithmSpecification().GetAlgorithmName() == flyteSageMakerIdl.AlgorithmName_CUSTOM {
-		return nil, errors.Errorf("Custom algorithm is not supported by the built-in training job plugin")
-	}
-
-	taskInput, err := taskCtx.InputReader().Get(ctx)
-	if err != nil {
-		return nil, errors.Wrapf(err, "unable to fetch task inputs")
-	}
-
-	// Get inputs from literals
-	inputLiterals := taskInput.GetLiterals()
-	err = m.checkIfRequiredInputLiteralsExist(inputLiterals,
-		[]string{TrainPredefinedInputVariable, ValidationPredefinedInputVariable, StaticHyperparametersPredefinedInputVariable})
-	if err != nil {
-		return nil, errors.Wrapf(err, "Error occurred when checking if all the required inputs exist")
-	}
-
-	trainPathLiteral := inputLiterals[TrainPredefinedInputVariable]
-	validationPathLiteral := inputLiterals[ValidationPredefinedInputVariable]
-	staticHyperparamsLiteral := inputLiterals[StaticHyperparametersPredefinedInputVariable]
-
-	if trainPathLiteral.GetScalar() == nil || trainPathLiteral.GetScalar().GetBlob() == nil {
-		return nil, errors.Errorf("[train] Input is required and should be of Type [Scalar.Blob]")
-	}
-	if validationPathLiteral.GetScalar() == nil || validationPathLiteral.GetScalar().GetBlob() == nil {
-		return nil, errors.Errorf("[validation] Input is required and should be of Type [Scalar.Blob]")
-	}
-
-	// Convert the hyperparameters to the spec value
-	staticHyperparams, err := convertStaticHyperparamsLiteralToSpecType(staticHyperparamsLiteral)
-	if err != nil {
-		return nil, errors.Wrapf(err, "could not convert static hyperparameters to spec type")
-	}
-
-	outputPath := createOutputPath(taskCtx.OutputWriter().GetRawOutputPrefix().String(), TrainingJobOutputPathSubDir)
-
-	jobName := taskCtx.TaskExecutionMetadata().GetTaskExecutionID().GetGeneratedName()
-
-	trainingImageStr, err := getTrainingJobImage(ctx, taskCtx, &sagemakerTrainingJob)
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to find the training image")
-	}
-
-	logger.Infof(ctx, "The Sagemaker TrainingJob Task plugin received static hyperparameters [%v]", staticHyperparams)
-
-	cfg := config.GetSagemakerConfig()
-
-	var metricDefinitions []commonv1.MetricDefinition
-	idlMetricDefinitions := sagemakerTrainingJob.GetAlgorithmSpecification().GetMetricDefinitions()
-	for _, md := range idlMetricDefinitions {
-		metricDefinitions = append(metricDefinitions,
-			commonv1.MetricDefinition{Name: ToStringPtr(md.Name), Regex: ToStringPtr(md.Regex)})
-	}
-
-	apiContentType, err := getAPIContentType(sagemakerTrainingJob.GetAlgorithmSpecification().GetInputContentType())
-	if err != nil {
-		return nil, errors.Wrapf(err, "Unsupported input file type [%v]", sagemakerTrainingJob.GetAlgorithmSpecification().GetInputContentType().String())
-	}
-
-	inputModeString := strings.Title(strings.ToLower(sagemakerTrainingJob.GetAlgorithmSpecification().GetInputMode().String()))
-
-	role := awsUtils.GetRole(ctx, cfg.RoleAnnotationKey, taskCtx.TaskExecutionMetadata().GetAnnotations())
-	if role == "" {
-		role = cfg.RoleArn
-	}
-
-	trainingJob := &trainingjobv1.TrainingJob{
-		Spec: trainingjobv1.TrainingJobSpec{
-			AlgorithmSpecification: &commonv1.AlgorithmSpecification{
-				// If the specify a value for this AlgorithmName parameter, the user can't specify a value for TrainingImage.
-				// in this Flyte plugin, we always use the algorithm name and version the user provides via Flytekit to map to an image
-				// so we intentionally leave this field nil
-				AlgorithmName:     nil,
-				TrainingImage:     ToStringPtr(trainingImageStr),
-				TrainingInputMode: commonv1.TrainingInputMode(inputModeString),
-				MetricDefinitions: metricDefinitions,
-			},
-			// The support of spot training will come in a later version
-			EnableManagedSpotTraining: nil,
-			HyperParameters:           staticHyperparams,
-			InputDataConfig: []commonv1.Channel{
-				{
-					ChannelName: ToStringPtr(TrainPredefinedInputVariable),
-					DataSource: &commonv1.DataSource{
-						S3DataSource: &commonv1.S3DataSource{
-							S3DataType: "S3Prefix",
-							S3Uri:      ToStringPtr(trainPathLiteral.GetScalar().GetBlob().GetUri()),
-						},
-					},
-					ContentType: ToStringPtr(apiContentType),
-					InputMode:   inputModeString,
-				},
-				{
-					ChannelName: ToStringPtr(ValidationPredefinedInputVariable),
-					DataSource: &commonv1.DataSource{
-						S3DataSource: &commonv1.S3DataSource{
-							S3DataType: "S3Prefix",
-							S3Uri:      ToStringPtr(validationPathLiteral.GetScalar().GetBlob().GetUri()),
-						},
-					},
-					ContentType: ToStringPtr(apiContentType),
-					InputMode:   inputModeString,
-				},
-			},
-			OutputDataConfig: &commonv1.OutputDataConfig{
-				S3OutputPath: ToStringPtr(outputPath),
-			},
-			CheckpointConfig: nil,
-			ResourceConfig: &commonv1.ResourceConfig{
-				InstanceType:   sagemakerTrainingJob.GetTrainingJobResourceConfig().GetInstanceType(),
-				InstanceCount:  ToInt64Ptr(sagemakerTrainingJob.GetTrainingJobResourceConfig().GetInstanceCount()),
-				VolumeSizeInGB: ToInt64Ptr(sagemakerTrainingJob.GetTrainingJobResourceConfig().GetVolumeSizeInGb()),
-				VolumeKmsKeyId: ToStringPtr(""), // TODO: Not yet supported. Need to add to proto and flytekit in the future
-			},
-			RoleArn: ToStringPtr(role),
-			Region:  ToStringPtr(cfg.Region),
-			StoppingCondition: &commonv1.StoppingCondition{
-				MaxRuntimeInSeconds:  ToInt64Ptr(86400), // TODO: decide how to coordinate this and Flyte's timeout
-				MaxWaitTimeInSeconds: nil,               // TODO: decide how to coordinate this and Flyte's timeout and queueing budget
-			},
-			TensorBoardOutputConfig: nil,
-			Tags:                    nil,
-			TrainingJobName:         &jobName,
-		},
-	}
-	logger.Infof(ctx, "Successfully built a training job resource for task [%v]", taskCtx.TaskExecutionMetadata().GetTaskExecutionID().GetGeneratedName())
-	return trainingJob, nil
-}
-
-func (m awsSagemakerPlugin) BuildResourceForCustomTrainingJob(
-	ctx context.Context, taskCtx pluginsCore.TaskExecutionContext) (k8s.Resource, error) {
-
-	logger.Infof(ctx, "Building a training job resource for task [%v]", taskCtx.TaskExecutionMetadata().GetTaskExecutionID().GetGeneratedName())
-	taskTemplate, err := getTaskTemplate(ctx, taskCtx)
-	if err != nil {
-		return nil, err
-	}
-
-	// Unmarshal the custom field of the task template back into the Hyperparameter Tuning Job struct generated in flyteidl
-	sagemakerTrainingJob := flyteSageMakerIdl.TrainingJob{}
-	err = utils.UnmarshalStruct(taskTemplate.GetCustom(), &sagemakerTrainingJob)
-	if err != nil {
-		return nil, errors.Wrapf(err, "invalid TrainingJob task specification: not able to unmarshal the custom field to [%s]", m.TaskType)
-	}
-
-	if sagemakerTrainingJob.GetAlgorithmSpecification() == nil {
-		return nil, errors.Errorf("The unmarshaled training job does not have a AlgorithmSpecification field")
-	}
-	if sagemakerTrainingJob.GetAlgorithmSpecification().GetAlgorithmName() != flyteSageMakerIdl.AlgorithmName_CUSTOM {
-		return nil, errors.Errorf("The algorithm name [%v] is not supported by the custom training job plugin",
-			sagemakerTrainingJob.GetAlgorithmSpecification().GetAlgorithmName().String())
-	}
-
-	inputChannels := make([]commonv1.Channel, 0)
-	inputModeString := strings.Title(strings.ToLower(sagemakerTrainingJob.GetAlgorithmSpecification().GetInputMode().String()))
-
-	jobName := taskCtx.TaskExecutionMetadata().GetTaskExecutionID().GetGeneratedName()
-	outputPath := taskCtx.OutputWriter().GetOutputPrefixPath().String()
-
-	if taskTemplate.GetContainer() == nil {
-		return nil, errors.Errorf("The task template points to a nil container")
-	}
-
-	if taskTemplate.GetContainer().GetImage() == "" {
-		return nil, errors.Errorf("Invalid image of the container")
-	}
-
-	trainingImageStr := taskTemplate.GetContainer().GetImage()
-
-	cfg := config.GetSagemakerConfig()
-
-	var metricDefinitions []commonv1.MetricDefinition
-	idlMetricDefinitions := sagemakerTrainingJob.GetAlgorithmSpecification().GetMetricDefinitions()
-	for _, md := range idlMetricDefinitions {
-		metricDefinitions = append(metricDefinitions,
-			commonv1.MetricDefinition{Name: ToStringPtr(md.Name), Regex: ToStringPtr(md.Regex)})
-	}
-
-	// TODO: When dealing with HPO job, we will need to deal with the following
-	// jobOutputPath := NewJobOutputPaths(ctx, taskCtx.DataStore(), taskCtx.OutputWriter().GetOutputPrefixPath(), jobName)
-
-	hyperParameters, err := injectArgsAndEnvVars(ctx, taskCtx, taskTemplate)
-	if err != nil {
-		return nil, errors.Wrapf(err, "Failed to inject the task template's container env vars to the hyperparameter list")
-	}
-
-	// Injecting a env var to disable statsd for sagemaker tasks
-	statsdDisableEnvVarName := fmt.Sprintf("%s%s%s", FlyteSageMakerEnvVarKeyPrefix, FlytesageMakerEnvVarKeyStatsdDisabled, FlyteSageMakerKeySuffix)
-	logger.Infof(ctx, "Injecting %v=%v to force disable statsd for SageMaker tasks only", statsdDisableEnvVarName, strconv.FormatBool(true))
-	hyperParameters = append(hyperParameters, &commonv1.KeyValuePair{
-		Name:  statsdDisableEnvVarName,
-		Value: strconv.FormatBool(true),
-	})
-
-	logger.Infof(ctx, "The Sagemaker TrainingJob Task plugin received static hyperparameters [%v]", hyperParameters)
-
-	trainingJob := &trainingjobv1.TrainingJob{
-		Spec: trainingjobv1.TrainingJobSpec{
-			AlgorithmSpecification: &commonv1.AlgorithmSpecification{
-				// If the specify a value for this AlgorithmName parameter, the user can't specify a value for TrainingImage.
-				// in this Flyte plugin, we always use the algorithm name and version the user provides via Flytekit to map to an image
-				// so we intentionally leave this field nil
-				AlgorithmName:     nil,
-				TrainingImage:     ToStringPtr(trainingImageStr),
-				TrainingInputMode: commonv1.TrainingInputMode(inputModeString),
-				MetricDefinitions: metricDefinitions,
-			},
-			// The support of spot training will come in a later version
-			EnableManagedSpotTraining: nil,
-			HyperParameters:           hyperParameters,
-			InputDataConfig:           inputChannels,
-			OutputDataConfig: &commonv1.OutputDataConfig{
-				S3OutputPath: ToStringPtr(outputPath),
-			},
-			CheckpointConfig: nil,
-			ResourceConfig: &commonv1.ResourceConfig{
-				InstanceType:   sagemakerTrainingJob.GetTrainingJobResourceConfig().GetInstanceType(),
-				InstanceCount:  ToInt64Ptr(sagemakerTrainingJob.GetTrainingJobResourceConfig().GetInstanceCount()),
-				VolumeSizeInGB: ToInt64Ptr(sagemakerTrainingJob.GetTrainingJobResourceConfig().GetVolumeSizeInGb()),
-				VolumeKmsKeyId: ToStringPtr(""), // TODO: Not yet supported. Need to add to proto and flytekit in the future
-			},
-			RoleArn: ToStringPtr(cfg.RoleArn),
-			Region:  ToStringPtr(cfg.Region),
-			StoppingCondition: &commonv1.StoppingCondition{
-				MaxRuntimeInSeconds:  ToInt64Ptr(86400), // TODO: decide how to coordinate this and Flyte's timeout
-				MaxWaitTimeInSeconds: nil,               // TODO: decide how to coordinate this and Flyte's timeout and queueing budget
-			},
-			TensorBoardOutputConfig: nil,
-			Tags:                    nil,
-			TrainingJobName:         &jobName,
-		},
-	}
-	logger.Infof(ctx, "Successfully built a custom training job resource for task [%v]", taskCtx.TaskExecutionMetadata().GetTaskExecutionID().GetGeneratedName())
-	return trainingJob, nil
-}
-
-func (m awsSagemakerPlugin) BuildResourceForHyperparameterTuningJob(
-	ctx context.Context, taskCtx pluginsCore.TaskExecutionContext) (k8s.Resource, error) {
-
-	logger.Infof(ctx, "Building a hyperparameter tuning job resource for task [%v]", taskCtx.TaskExecutionMetadata().GetTaskExecutionID().GetGeneratedName())
-
-	taskTemplate, err := getTaskTemplate(ctx, taskCtx)
-	if err != nil {
-		return nil, err
-	}
-
-	// Unmarshal the custom field of the task template back into the HyperparameterTuningJob struct generated in flyteidl
-	sagemakerHPOJob := flyteSageMakerIdl.HyperparameterTuningJob{}
-	err = utils.UnmarshalStruct(taskTemplate.GetCustom(), &sagemakerHPOJob)
-	if err != nil {
-		return nil, errors.Wrapf(err, "invalid HyperparameterTuningJob task specification: not able to unmarshal the custom field to [%s]", hyperparameterTuningJobTaskType)
-	}
-	if sagemakerHPOJob.GetTrainingJob() == nil {
-		return nil, errors.Errorf("Required field [TrainingJob] of the HyperparameterTuningJob does not exist")
-	}
-	if sagemakerHPOJob.GetTrainingJob().GetAlgorithmSpecification() == nil {
-		return nil, errors.Errorf("Required field [AlgorithmSpecification] of the HyperparameterTuningJob's underlying TrainingJob does not exist")
-	}
-	if sagemakerHPOJob.GetTrainingJob().GetTrainingJobResourceConfig() == nil {
-		return nil, errors.Errorf("Required field [TrainingJobResourceConfig] of the HyperparameterTuningJob's underlying TrainingJob does not exist")
-	}
-
-	taskInput, err := taskCtx.InputReader().Get(ctx)
-	if err != nil {
-		return nil, errors.Wrapf(err, "unable to fetch task inputs")
-	}
-
-	// Get inputs from literals
-	inputLiterals := taskInput.GetLiterals()
-	err = m.checkIfRequiredInputLiteralsExist(inputLiterals,
-		[]string{TrainPredefinedInputVariable, ValidationPredefinedInputVariable, StaticHyperparametersPredefinedInputVariable})
-	if err != nil {
-		return nil, errors.Wrapf(err, "Error occurred when checking if all the required inputs exist")
-	}
-
-	trainPathLiteral := inputLiterals[TrainPredefinedInputVariable]
-	validatePathLiteral := inputLiterals[ValidationPredefinedInputVariable]
-	staticHyperparamsLiteral := inputLiterals[StaticHyperparametersPredefinedInputVariable]
-	hpoJobConfigLiteral := inputLiterals["hyperparameter_tuning_job_config"]
-	if trainPathLiteral.GetScalar() == nil || trainPathLiteral.GetScalar().GetBlob() == nil {
-		return nil, errors.Errorf("[%v] Input is required and should be of Type [Scalar.Blob]", TrainPredefinedInputVariable)
-	}
-	if validatePathLiteral.GetScalar() == nil || validatePathLiteral.GetScalar().GetBlob() == nil {
-		return nil, errors.Errorf("[%v] Input is required and should be of Type [Scalar.Blob]", ValidationPredefinedInputVariable)
-	}
-	// Convert the hyperparameters to the spec value
-	staticHyperparams, err := convertStaticHyperparamsLiteralToSpecType(staticHyperparamsLiteral)
-	if err != nil {
-		return nil, errors.Wrapf(err, "could not convert static hyperparameters to spec type")
-	}
-	// hyperparameter_tuning_job_config is marshaled into a byte array in flytekit, so will have to unmarshal it back
-	hpoJobConfig, err := convertHyperparameterTuningJobConfigToSpecType(hpoJobConfigLiteral)
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to convert hyperparameter tuning job config literal to spec type")
-	}
-
-	outputPath := createOutputPath(taskCtx.OutputWriter().GetRawOutputPrefix().String(), HyperparameterOutputPathSubDir)
-
-	if hpoJobConfig.GetTuningObjective() == nil {
-		return nil, errors.Errorf("Required field [TuningObjective] does not exist")
-	}
-
-	// Deleting the conflicting static hyperparameters: if a hyperparameter exist in both the map of static hyperparameter
-	// and the map of the tunable hyperparameter inside the Hyperparameter Tuning Job Config, we delete the entry
-	// in the static map and let the one in the map of the tunable hyperparameters take precedence
-	staticHyperparams = deleteConflictingStaticHyperparameters(ctx, staticHyperparams, hpoJobConfig.GetHyperparameterRanges().GetParameterRangeMap())
-
-	jobName := taskCtx.TaskExecutionMetadata().GetTaskExecutionID().GetGeneratedName()
-
-	trainingImageStr, err := getTrainingJobImage(ctx, taskCtx, sagemakerHPOJob.GetTrainingJob())
-	if err != nil {
-		return nil, errors.Wrapf(err, "failed to find the training image")
-	}
-
-	hpoJobParameterRanges := buildParameterRanges(hpoJobConfig)
-
-	logger.Infof(ctx, "The Sagemaker HyperparameterTuningJob Task plugin received the following inputs: \n"+
-		"static hyperparameters: [%v]\n"+
-		"hyperparameter tuning job config: [%v]\n"+
-		"parameter ranges: [%v]", staticHyperparams, hpoJobConfig, hpoJobParameterRanges)
-
-	cfg := config.GetSagemakerConfig()
-
-	var metricDefinitions []commonv1.MetricDefinition
-	idlMetricDefinitions := sagemakerHPOJob.GetTrainingJob().GetAlgorithmSpecification().GetMetricDefinitions()
-	for _, md := range idlMetricDefinitions {
-		metricDefinitions = append(metricDefinitions,
-			commonv1.MetricDefinition{Name: ToStringPtr(md.Name), Regex: ToStringPtr(md.Regex)})
-	}
-
-	apiContentType, err := getAPIContentType(sagemakerHPOJob.GetTrainingJob().GetAlgorithmSpecification().GetInputContentType())
-	if err != nil {
-		return nil, errors.Wrapf(err, "Unsupported input file type [%v]",
-			sagemakerHPOJob.GetTrainingJob().GetAlgorithmSpecification().GetInputContentType().String())
-	}
-
-	inputModeString := strings.Title(strings.ToLower(sagemakerHPOJob.GetTrainingJob().GetAlgorithmSpecification().GetInputMode().String()))
-	tuningStrategyString := strings.Title(strings.ToLower(hpoJobConfig.GetTuningStrategy().String()))
-	tuningObjectiveTypeString := strings.Title(strings.ToLower(hpoJobConfig.GetTuningObjective().GetObjectiveType().String()))
-	trainingJobEarlyStoppingTypeString := strings.Title(strings.ToLower(hpoJobConfig.TrainingJobEarlyStoppingType.String()))
-
-	role := awsUtils.GetRole(ctx, cfg.RoleAnnotationKey, taskCtx.TaskExecutionMetadata().GetAnnotations())
-	if role == "" {
-		role = cfg.RoleArn
-	}
-
-	hpoJob := &hpojobv1.HyperparameterTuningJob{
-		Spec: hpojobv1.HyperparameterTuningJobSpec{
-			HyperParameterTuningJobName: &jobName,
-			HyperParameterTuningJobConfig: &commonv1.HyperParameterTuningJobConfig{
-				ResourceLimits: &commonv1.ResourceLimits{
-					MaxNumberOfTrainingJobs: ToInt64Ptr(sagemakerHPOJob.GetMaxNumberOfTrainingJobs()),
-					MaxParallelTrainingJobs: ToInt64Ptr(sagemakerHPOJob.GetMaxParallelTrainingJobs()),
-				},
-				Strategy: commonv1.HyperParameterTuningJobStrategyType(tuningStrategyString),
-				HyperParameterTuningJobObjective: &commonv1.HyperParameterTuningJobObjective{
-					Type:       commonv1.HyperParameterTuningJobObjectiveType(tuningObjectiveTypeString),
-					MetricName: ToStringPtr(hpoJobConfig.GetTuningObjective().GetMetricName()),
-				},
-				ParameterRanges:              hpoJobParameterRanges,
-				TrainingJobEarlyStoppingType: commonv1.TrainingJobEarlyStoppingType(trainingJobEarlyStoppingTypeString),
-			},
-			TrainingJobDefinition: &commonv1.HyperParameterTrainingJobDefinition{
-				StaticHyperParameters: staticHyperparams,
-				AlgorithmSpecification: &commonv1.HyperParameterAlgorithmSpecification{
-					TrainingImage:     ToStringPtr(trainingImageStr),
-					TrainingInputMode: commonv1.TrainingInputMode(inputModeString),
-					MetricDefinitions: metricDefinitions,
-					AlgorithmName:     nil,
-				},
-				InputDataConfig: []commonv1.Channel{
-					{
-						ChannelName: ToStringPtr(TrainPredefinedInputVariable),
-						DataSource: &commonv1.DataSource{
-							S3DataSource: &commonv1.S3DataSource{
-								S3DataType: "S3Prefix",
-								S3Uri:      ToStringPtr(trainPathLiteral.GetScalar().GetBlob().GetUri()),
-							},
-						},
-						ContentType: ToStringPtr(apiContentType), // TODO: can this be derived from the BlobMetadata
-						InputMode:   inputModeString,
-					},
-					{
-						ChannelName: ToStringPtr(ValidationPredefinedInputVariable),
-						DataSource: &commonv1.DataSource{
-							S3DataSource: &commonv1.S3DataSource{
-								S3DataType: "S3Prefix",
-								S3Uri:      ToStringPtr(validatePathLiteral.GetScalar().GetBlob().GetUri()),
-							},
-						},
-						ContentType: ToStringPtr(apiContentType), // TODO: can this be derived from the BlobMetadata
-						InputMode:   inputModeString,
-					},
-				},
-				OutputDataConfig: &commonv1.OutputDataConfig{
-					S3OutputPath: ToStringPtr(outputPath),
-				},
-				ResourceConfig: &commonv1.ResourceConfig{
-					InstanceType:   sagemakerHPOJob.GetTrainingJob().GetTrainingJobResourceConfig().GetInstanceType(),
-					InstanceCount:  ToInt64Ptr(sagemakerHPOJob.GetTrainingJob().GetTrainingJobResourceConfig().GetInstanceCount()),
-					VolumeSizeInGB: ToInt64Ptr(sagemakerHPOJob.GetTrainingJob().GetTrainingJobResourceConfig().GetVolumeSizeInGb()),
-					VolumeKmsKeyId: ToStringPtr(""), // TODO: Not yet supported. Need to add to proto and flytekit in the future
-				},
-				RoleArn: ToStringPtr(role),
-				StoppingCondition: &commonv1.StoppingCondition{
-					MaxRuntimeInSeconds:  ToInt64Ptr(86400),
-					MaxWaitTimeInSeconds: nil,
-				},
-			},
-			Region: ToStringPtr(cfg.Region),
-		},
-	}
-
-	logger.Infof(ctx, "Successfully built a hyperparameter tuning job resource for task [%v]", taskCtx.TaskExecutionMetadata().GetTaskExecutionID().GetGeneratedName())
-	return hpoJob, nil
-}
-
-func getTaskTemplate(ctx context.Context, taskCtx pluginsCore.TaskExecutionContext) (*flyteIdlCore.TaskTemplate, error) {
-	taskTemplate, err := taskCtx.TaskReader().Read(ctx)
-	if err != nil {
-		return nil, errors.Wrapf(err, "unable to fetch task specification")
-	} else if taskTemplate == nil {
-		return nil, errors.Errorf("nil task specification")
-	}
-	return taskTemplate, nil
-}
-
-func (m awsSagemakerPlugin) BuildResource(ctx context.Context, taskCtx pluginsCore.TaskExecutionContext) (k8s.Resource, error) {
-
-	// Unmarshal the custom field of the task template back into the HyperparameterTuningJob struct generated in flyteidl
-	if m.TaskType == trainingJobTaskType {
-		return m.BuildResourceForTrainingJob(ctx, taskCtx)
-	}
-	if m.TaskType == customTrainingJobTaskType {
-		return m.BuildResourceForCustomTrainingJob(ctx, taskCtx)
-	}
-	if m.TaskType == hyperparameterTuningJobTaskType {
-		return m.BuildResourceForHyperparameterTuningJob(ctx, taskCtx)
-	}
-	return nil, errors.Errorf("The SageMaker plugin is unable to build resource for unknown task type [%s]", m.TaskType)
-}
-
-func (m awsSagemakerPlugin) getEventInfoForJob(ctx context.Context, job k8s.Resource) (*pluginsCore.TaskInfo, error) {
-
-	var jobRegion, jobName, jobTypeInURL, sagemakerLinkName string
-	if m.TaskType == trainingJobTaskType {
-		trainingJob := job.(*trainingjobv1.TrainingJob)
-		jobRegion = *trainingJob.Spec.Region
-		jobName = *trainingJob.Spec.TrainingJobName
-		jobTypeInURL = "jobs"
-		sagemakerLinkName = "SageMaker Training Job"
-	} else if m.TaskType == customTrainingJobTaskType {
-		trainingJob := job.(*trainingjobv1.TrainingJob)
-		jobRegion = *trainingJob.Spec.Region
-		jobName = *trainingJob.Spec.TrainingJobName
-		jobTypeInURL = "jobs"
-		sagemakerLinkName = "SageMaker Custom Training Job"
-	} else if m.TaskType == hyperparameterTuningJobTaskType {
-		trainingJob := job.(*hpojobv1.HyperparameterTuningJob)
-		jobRegion = *trainingJob.Spec.Region
-		jobName = *trainingJob.Spec.HyperParameterTuningJobName
-		jobTypeInURL = "hyper-tuning-jobs"
-		sagemakerLinkName = "SageMaker Hyperparameter Tuning Job"
-	} else {
-		return nil, errors.Errorf("The plugin is unable to get event info for unknown task type {%v}", m.TaskType)
-	}
-
-	logger.Infof(ctx, "Getting event information for task type: [%v], job region: [%v], job name: [%v], "+
-		"job type in url: [%v], sagemaker link name: [%v]", m.TaskType, jobRegion, jobName, jobTypeInURL, sagemakerLinkName)
-
-	cwLogURL := fmt.Sprintf("https://%s.console.aws.amazon.com/cloudwatch/home?region=%s#logStream:group=/aws/sagemaker/TrainingJobs;prefix=%s;streamFilter=typeLogStreamPrefix",
-		jobRegion, jobRegion, jobName)
-	smLogURL := fmt.Sprintf("https://%s.console.aws.amazon.com/sagemaker/home?region=%s#/%s/%s",
-		jobRegion, jobRegion, jobTypeInURL, jobName)
-
-	taskLogs := []*flyteIdlCore.TaskLog{
-		{
-			Uri:           cwLogURL,
-			Name:          "CloudWatch Logs",
-			MessageFormat: flyteIdlCore.TaskLog_JSON,
-		},
-		{
-			Uri:           smLogURL,
-			Name:          sagemakerLinkName,
-			MessageFormat: flyteIdlCore.TaskLog_UNKNOWN,
-		},
-	}
-
-	customInfoMap := make(map[string]string)
-
-	customInfo, err := utils.MarshalObjToStruct(customInfoMap)
-	if err != nil {
-		return nil, err
-	}
-
-	return &pluginsCore.TaskInfo{
-		Logs:       taskLogs,
-		CustomInfo: customInfo,
-	}, nil
-}
-
-func getOutputLiteralMapFromTaskInterface(ctx context.Context, tr pluginsCore.TaskReader, outputPath string) (*flyteIdlCore.LiteralMap, error) {
-	tk, err := tr.Read(ctx)
-	if err != nil {
-		return nil, err
-	}
-	if tk.Interface.Outputs != nil && tk.Interface.Outputs.Variables == nil {
-		logger.Warnf(ctx, "No outputs declared in the output interface. Ignoring the generated outputs.")
-		return nil, nil
-	}
-
-	// We know that for XGBoost task there is only one output to be generated
-	if len(tk.Interface.Outputs.Variables) > 1 {
-		return nil, fmt.Errorf("expected to generate more than one outputs of type [%v]", tk.Interface.Outputs.Variables)
-	}
-	op := createOutputLiteralMap(tk, outputPath)
-	return op, nil
-}
-
-func createOutputPath(prefix string, subdir string) string {
-	return fmt.Sprintf("%s/%s", prefix, subdir)
-}
-
-func createModelOutputPath(job k8s.Resource, prefix, jobName string) string {
-	switch job.(type) {
-	case *trainingjobv1.TrainingJob:
-		return fmt.Sprintf("%s/%s/output/model.tar.gz", createOutputPath(prefix, TrainingJobOutputPathSubDir), jobName)
-	case *hpojobv1.HyperparameterTuningJob:
-		return fmt.Sprintf("%s/%s/output/model.tar.gz", createOutputPath(prefix, HyperparameterOutputPathSubDir), jobName)
-	default:
-		return fmt.Sprintf("")
-	}
-}
-
-func (m awsSagemakerPlugin) GetTaskPhaseForTrainingJob(
-	ctx context.Context, pluginContext k8s.PluginContext, trainingJob *trainingjobv1.TrainingJob) (pluginsCore.PhaseInfo, error) {
-
-	logger.Infof(ctx, "Getting task phase for sagemaker training job [%v]", trainingJob.Status.SageMakerTrainingJobName)
-	info, err := m.getEventInfoForJob(ctx, trainingJob)
-	if err != nil {
-		return pluginsCore.PhaseInfoUndefined, err
-	}
-
-	occurredAt := time.Now()
-
-	switch trainingJob.Status.TrainingJobStatus {
-	case trainingjobController.ReconcilingTrainingJobStatus:
-		logger.Errorf(ctx, "Job stuck in reconciling status, assuming retryable failure [%s]", trainingJob.Status.Additional)
-		// TODO talk to AWS about why there cannot be an explicit condition that signals AWS API call errors
-		execError := &flyteIdlCore.ExecutionError{
-			Message: trainingJob.Status.Additional,
-			Kind:    flyteIdlCore.ExecutionError_USER,
-			Code:    trainingjobController.ReconcilingTrainingJobStatus,
-		}
-		return pluginsCore.PhaseInfoFailed(pluginsCore.PhaseRetryableFailure, execError, info), nil
-	case sagemaker.TrainingJobStatusFailed:
-		execError := &flyteIdlCore.ExecutionError{
-			Message: trainingJob.Status.Additional,
-			Kind:    flyteIdlCore.ExecutionError_USER,
-			Code:    sagemaker.TrainingJobStatusFailed,
-		}
-		return pluginsCore.PhaseInfoFailed(pluginsCore.PhasePermanentFailure, execError, info), nil
-	case sagemaker.TrainingJobStatusStopped:
-		reason := fmt.Sprintf("Training Job Stopped")
-		return pluginsCore.PhaseInfoRetryableFailure(taskError.DownstreamSystemError, reason, info), nil
-	case sagemaker.TrainingJobStatusCompleted:
-		// Now that it is a success we will set the outputs as expected by the task
-
-		// We have specified an output path in the CRD, and we know SageMaker will automatically upload the
-		// model tarball to s3://<specified-output-path>/<training-job-name>/output/model.tar.gz
-
-		// Therefore, here we create a output literal map, where we fill in the above path to the URI field of the
-		// blob output, which will later be written out by the OutputWriter to the outputs.pb remotely on S3
-		outputLiteralMap, err := getOutputLiteralMapFromTaskInterface(ctx, pluginContext.TaskReader(),
-			createModelOutputPath(trainingJob, pluginContext.OutputWriter().GetRawOutputPrefix().String(), trainingJob.Status.SageMakerTrainingJobName))
-		if err != nil {
-			logger.Errorf(ctx, "Failed to create outputs, err: %s", err)
-			return pluginsCore.PhaseInfoUndefined, errors.Wrapf(err, "failed to create outputs for the task")
-		}
-		// Instantiate a output reader with the literal map, and write the output to the remote location referred to by the OutputWriter
-		if err := pluginContext.OutputWriter().Put(ctx, ioutils.NewInMemoryOutputReader(outputLiteralMap, nil)); err != nil {
-			return pluginsCore.PhaseInfoUndefined, err
-		}
-		logger.Debugf(ctx, "Successfully produced and returned outputs")
-		return pluginsCore.PhaseInfoSuccess(info), nil
-	case "":
-		return pluginsCore.PhaseInfoQueued(occurredAt, pluginsCore.DefaultPhaseVersion, "job submitted"), nil
-	}
-
-	return pluginsCore.PhaseInfoRunning(pluginsCore.DefaultPhaseVersion, info), nil
-}
-
-func (m awsSagemakerPlugin) GetTaskPhaseForCustomTrainingJob(
-	ctx context.Context, pluginContext k8s.PluginContext, trainingJob *trainingjobv1.TrainingJob) (pluginsCore.PhaseInfo, error) {
-
-	logger.Infof(ctx, "Getting task phase for sagemaker training job [%v]", trainingJob.Status.SageMakerTrainingJobName)
-	info, err := m.getEventInfoForJob(ctx, trainingJob)
-	if err != nil {
-		return pluginsCore.PhaseInfoUndefined, err
-	}
-
-	occurredAt := time.Now()
-
-	switch trainingJob.Status.TrainingJobStatus {
-	case trainingjobController.ReconcilingTrainingJobStatus:
-		logger.Errorf(ctx, "Job stuck in reconciling status, assuming retryable failure [%s]", trainingJob.Status.Additional)
-		// TODO talk to AWS about why there cannot be an explicit condition that signals AWS API call errors
-		execError := &flyteIdlCore.ExecutionError{
-			Message: trainingJob.Status.Additional,
-			Kind:    flyteIdlCore.ExecutionError_USER,
-			Code:    trainingjobController.ReconcilingTrainingJobStatus,
-		}
-		return pluginsCore.PhaseInfoFailed(pluginsCore.PhaseRetryableFailure, execError, info), nil
-	case sagemaker.TrainingJobStatusFailed:
-		execError := &flyteIdlCore.ExecutionError{
-			Message: trainingJob.Status.Additional,
-			Kind:    flyteIdlCore.ExecutionError_USER,
-			Code:    sagemaker.TrainingJobStatusFailed,
-		}
-		return pluginsCore.PhaseInfoFailed(pluginsCore.PhasePermanentFailure, execError, info), nil
-	case sagemaker.TrainingJobStatusStopped:
-		reason := fmt.Sprintf("Training Job Stopped")
-		return pluginsCore.PhaseInfoRetryableFailure(taskError.DownstreamSystemError, reason, info), nil
-	case sagemaker.TrainingJobStatusCompleted:
-		// Now that it is a success we will set the outputs as expected by the task
-
-		logger.Infof(ctx, "Looking for the output.pb under %s", pluginContext.OutputWriter().GetOutputPrefixPath())
-		outputReader := ioutils.NewRemoteFileOutputReader(ctx, pluginContext.DataStore(), pluginContext.OutputWriter(), pluginContext.MaxDatasetSizeBytes())
-
-		// Instantiate a output reader with the literal map, and write the output to the remote location referred to by the OutputWriter
-		if err := pluginContext.OutputWriter().Put(ctx, outputReader); err != nil {
-			return pluginsCore.PhaseInfoUndefined, err
-		}
-		logger.Debugf(ctx, "Successfully produced and returned outputs")
-		return pluginsCore.PhaseInfoSuccess(info), nil
-	case "":
-		return pluginsCore.PhaseInfoQueued(occurredAt, pluginsCore.DefaultPhaseVersion, "job submitted"), nil
-	}
-
-	return pluginsCore.PhaseInfoRunning(pluginsCore.DefaultPhaseVersion, info), nil
-}
-
-func (m awsSagemakerPlugin) GetTaskPhaseForHyperparameterTuningJob(
-	ctx context.Context, pluginContext k8s.PluginContext, hpoJob *hpojobv1.HyperparameterTuningJob) (pluginsCore.PhaseInfo, error) {
-
-	logger.Infof(ctx, "Getting task phase for hyperparameter tuning job [%v]", hpoJob.Status.SageMakerHyperParameterTuningJobName)
-	info, err := m.getEventInfoForJob(ctx, hpoJob)
-	if err != nil {
-		return pluginsCore.PhaseInfoUndefined, err
-	}
-
-	occurredAt := time.Now()
-
-	switch hpoJob.Status.HyperParameterTuningJobStatus {
-	case hpojobController.ReconcilingTuningJobStatus:
-		logger.Errorf(ctx, "Job stuck in reconciling status, assuming retryable failure [%s]", hpoJob.Status.Additional)
-		// TODO talk to AWS about why there cannot be an explicit condition that signals AWS API call errors
-		execError := &flyteIdlCore.ExecutionError{
-			Message: hpoJob.Status.Additional,
-			Kind:    flyteIdlCore.ExecutionError_USER,
-			Code:    hpojobController.ReconcilingTuningJobStatus,
-		}
-		return pluginsCore.PhaseInfoFailed(pluginsCore.PhaseRetryableFailure, execError, info), nil
-	case sagemaker.HyperParameterTuningJobStatusFailed:
-		execError := &flyteIdlCore.ExecutionError{
-			Message: hpoJob.Status.Additional,
-			Kind:    flyteIdlCore.ExecutionError_USER,
-			Code:    sagemaker.HyperParameterTuningJobStatusFailed,
-		}
-		return pluginsCore.PhaseInfoFailed(pluginsCore.PhasePermanentFailure, execError, info), nil
-	case sagemaker.HyperParameterTuningJobStatusStopped:
-		reason := fmt.Sprintf("Hyperparameter tuning job stopped")
-		return pluginsCore.PhaseInfoRetryableFailure(taskError.DownstreamSystemError, reason, info), nil
-	case sagemaker.HyperParameterTuningJobStatusCompleted:
-		// Now that it is a success we will set the outputs as expected by the task
-
-		// TODO:
-		// Check task template -> custom training job -> if custom: assume output.pb exist, and fail if it doesn't. If it exists, then
-		//						 				      -> if not custom: check model.tar.gz
-		out, err := getOutputLiteralMapFromTaskInterface(ctx, pluginContext.TaskReader(),
-			createModelOutputPath(hpoJob, pluginContext.OutputWriter().GetRawOutputPrefix().String(),
-				*hpoJob.Status.BestTrainingJob.TrainingJobName))
-		if err != nil {
-			logger.Errorf(ctx, "Failed to create outputs, err: %s", err)
-			return pluginsCore.PhaseInfoUndefined, errors.Wrapf(err, "failed to create outputs for the task")
-		}
-		if err := pluginContext.OutputWriter().Put(ctx, ioutils.NewInMemoryOutputReader(out, nil)); err != nil {
-			return pluginsCore.PhaseInfoUndefined, err
-		}
-		logger.Debugf(ctx, "Successfully produced and returned outputs")
-		return pluginsCore.PhaseInfoSuccess(info), nil
-	case "":
-		return pluginsCore.PhaseInfoQueued(occurredAt, pluginsCore.DefaultPhaseVersion, "job submitted"), nil
-	}
-
-	return pluginsCore.PhaseInfoRunning(pluginsCore.DefaultPhaseVersion, info), nil
-}
-
-func (m awsSagemakerPlugin) GetTaskPhase(ctx context.Context, pluginContext k8s.PluginContext, resource k8s.Resource) (pluginsCore.PhaseInfo, error) {
-	if m.TaskType == trainingJobTaskType {
-		job := resource.(*trainingjobv1.TrainingJob)
-		return m.GetTaskPhaseForTrainingJob(ctx, pluginContext, job)
-	} else if m.TaskType == customTrainingJobTaskType {
-		job := resource.(*trainingjobv1.TrainingJob)
-		return m.GetTaskPhaseForCustomTrainingJob(ctx, pluginContext, job)
-	} else if m.TaskType == hyperparameterTuningJobTaskType {
-		job := resource.(*hpojobv1.HyperparameterTuningJob)
-		return m.GetTaskPhaseForHyperparameterTuningJob(ctx, pluginContext, job)
-	}
-	return pluginsCore.PhaseInfoUndefined, errors.Errorf("cannot get task phase for unknown task type [%s]", m.TaskType)
-}
-
-func init() {
-	if err := commonv1.AddToScheme(scheme.Scheme); err != nil {
-		panic(err)
-	}
-
-	// Registering the plugin for HyperparameterTuningJob
-	pluginmachinery.PluginRegistry().RegisterK8sPlugin(
-		k8s.PluginEntry{
-			ID:                  hyperparameterTuningJobTaskPluginID,
-			RegisteredTaskTypes: []pluginsCore.TaskType{hyperparameterTuningJobTaskType},
-			ResourceToWatch:     &hpojobv1.HyperparameterTuningJob{},
-			Plugin:              awsSagemakerPlugin{TaskType: hyperparameterTuningJobTaskType},
-			IsDefault:           false,
-		})
-
-	// Registering the plugin for standalone TrainingJob
-	pluginmachinery.PluginRegistry().RegisterK8sPlugin(
-		k8s.PluginEntry{
-			ID:                  trainingJobTaskPluginID,
-			RegisteredTaskTypes: []pluginsCore.TaskType{trainingJobTaskType},
-			ResourceToWatch:     &trainingjobv1.TrainingJob{},
-			Plugin:              awsSagemakerPlugin{TaskType: trainingJobTaskType},
-			IsDefault:           false,
-		})
-
-	// Registering the plugin for custom TrainingJob
-	pluginmachinery.PluginRegistry().RegisterK8sPlugin(
-		k8s.PluginEntry{
-			ID:                  customTrainingJobTaskPluginID,
-			RegisteredTaskTypes: []pluginsCore.TaskType{customTrainingJobTaskType},
-			ResourceToWatch:     &trainingjobv1.TrainingJob{},
-			Plugin:              awsSagemakerPlugin{TaskType: customTrainingJobTaskType},
-			IsDefault:           false,
-		})
-}
